Cleaning up thirdweb createWallet() methods, in a single palce
