+++ components/TransactionPreview/TransactionActions.tsx
@@ -1,4 +1,3 @@
-import { InstalledWallet } from "@components/Onboarding/ConnectViaWallet/ConnectViaWalletSupportedWallets";
 import { Text } from "@design-system/Text";
 import { TouchableOpacity } from "@design-system/TouchableOpacity";
 import { VStack } from "@design-system/VStack";
@@ -7,7 +6,7 @@ import { spacing } from "@theme/spacing";
 
 import { TransactionToTrigger } from "./TransactionPreview";
 import TransactionNext from "../../assets/transaction-next.svg";
-import { CHAIN_BY_ID } from "@utils/evm/wallets";
+import { CHAIN_BY_ID, InstalledWallet } from "@utils/evm/wallets";
 
 type ITransactionActionsProps = {
   shouldSwitchChain?: boolean;
