+++ utils/thirdweb.ts
@@ -1,6 +1,8 @@
 import { createThirdwebClient } from "thirdweb";
 
 import config from "../config";
+import { createWallet, Wallet, WalletId } from "thirdweb/wallets";
+import { ISupportedWalletName, SUPPORTED_WALLETS } from "./evm/wallets";
 
 const thirdwebGateway = `https://${config.thirdwebClientId}.ipfscdn.io`;
 
@@ -18,3 +20,40 @@ export const getIPFSAssetURI = (ipfsURI?: string) => {
 export const thirdwebClient = createThirdwebClient({
   clientId: config.thirdwebClientId,
 });
+
+export const thirdwebWallets: Record<ISupportedWalletName, Wallet> =
+  Object.fromEntries(
+    Object.entries(SUPPORTED_WALLETS).map(([walletName, walletConfig]) => {
+      if (!walletConfig.thirdwebId) {
+        return [walletName, undefined];
+      }
+
+      let wallet: Wallet;
+
+      if (walletName === "Coinbase Smart Wallet") {
+        wallet = createWallet("com.coinbase.wallet", {
+          appMetadata: config.walletConnectConfig.appMetadata,
+          mobileConfig: {
+            callbackURL: `converse-dev://mobile-wallet-protocol`,
+          },
+          walletConfig: {
+            options: "smartWalletOnly",
+          },
+        });
+      } else if (walletName === "Coinbase Wallet") {
+        wallet = createWallet("com.coinbase.wallet", {
+          appMetadata: config.walletConnectConfig.appMetadata,
+          mobileConfig: {
+            callbackURL: `https://${config.websiteDomain}/coinbase`,
+          },
+          walletConfig: {
+            options: "eoaOnly",
+          },
+        });
+      } else {
+        wallet = createWallet(walletConfig.thirdwebId);
+      }
+
+      return [walletName, wallet];
+    })
+  ) as Record<ISupportedWalletName, Wallet>;
